+++ changelog.md
@@ -1,5 +1,15 @@
 # Changelog
 
+## [Unreleased]
+
+### Added
+
+- Added Plain as third method, which does not encrypt or decrypt anything
+
+### Changed
+
+- Log another possible Sodium error
+
 ## [3.0.2] - 17.08.2026
 
 ### Fixed
@@ -21,7 +31,7 @@
 - Added error if no usable place could be found
 - Added support for wp-config.php in the parent directory of the WordPress root (as wp-load.php does)
 - Added multiple more PHP Unit Test to ensure the functionality of this package
-- Added documentation for the hooks this package provides
+- Added a documentation for the hooks this package provides
 
 ### Changed